Remove jquery and shiny JS/CSS imports
These will be brought in by Shiny itself.

It wasn't our intention to break compatibility with existing .html apps, and so I can't promise that we won't go backwards on this once we fully investigate (i.e. this change may need to be reversed in the near future once we fully understand what happened).